pyproject.toml
[project]
name = "fhir-x-synthea-csv"
version = "0.1.0"
description = "Bidirectional mapping library between Synthea CSV and FHIR R4B resources"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chidian>=0.2.1",
    "fhir.resources>=8.1.0",
    "synthea-pydantic>=0.1.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "pytest-cov>=6.2.1",
    "ruff>=0.12.9",
]
